Restructured + added makefile
