+++ src/internal/QueryInterface/MasterEngine.go
@@ -1,7 +1,7 @@
 package QueryInterface
 
 import (
-	"KVS/Storage"
+	"KVS/src/internal/Storage"
 	"sync"
 )
