+++ src/internal/Storage/LinkedList.go
