+++ src/internal/Storage/storageFunc.go
